ISSUE #4909 - 3d calibration handler is unmounted when changing step
